requirements.txt
fastapi
uvicorn[standard]
python-dotenv
torch --index-url https://download.pytorch.org/whl/cu121
vllm
sentence-transformers
deepgram-sdk